+++ src/Home.js
@@ -4,19 +4,24 @@ import Box from '@mui/material/Box';
 import Grid from '@mui/material/Grid2';
 import MainSection from './MainSection';
 import Divider from '@mui/material/Divider';
+import { useMediaQuery } from '@mui/material';
 
 
 
 const Home = () =>{
+
+    const isSmallScreen = useMediaQuery('(max-width:1200px)');
       
     return(
         <>
-            <Box sx={{ flexGrow: 1}}>
+            <Box sx={{ flexGrow: 1,m:1}}>
             <Grid container sx={{height:"100vh"}}>
-                <Grid size={2}>
+            {isSmallScreen? null : 
+                    <Grid size={2}>
                     <SideNavBar/>
                 </Grid>
-                <Grid size={10}>
+            }
+                <Grid size={{sm:12,md:12,lg:10}}>
                     <MainSection/>
                 </Grid>
             </Grid>
